refactor(connection): connection as namespace allow for better readability of the module
