+++ lib/connection/unreachable.rb
@@ -1,10 +1,12 @@
-require_relative './connection'
+require_relative './base'
 require_relative '../errors'
 
 module Ipfs
-  class Unreachable < Connection
-    def up?
-      raise Ipfs::Error::UnreachableDaemon, "IPFS is not reachable."
+  module Connection
+    class Unreachable < Base
+      def up?
+        raise Ipfs::Error::UnreachableDaemon, "IPFS is not reachable."
+      end
     end
   end
 end
